Major update to package.json
It now shows the configuration settings under Enmeti as it should, so if you go to File → Preferences → Settings → Extensions, you'll find Enmeti in the list.

I've updated the devDependencies to the latest versions available.

I also updated the version number to 1.1.1 as well as the display name to no longer include (MD Link Inserter) since it does more than just insert links at this point.